V2_curso/django_project/pyproject.toml
[tool.poetry]
name = "django-project"
version = "0.1.0"
description = "Bookstore API"
authors = ["Daniel Assuncao <assuncao.daniel89@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
djangorestframework = "^3.14.0"
markdown = "^3.5.1"
django-filter = "^23.5"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
